+++ src/operations/sync.ts
@@ -5,8 +5,9 @@ import Namespaces from '../resources/namespaces/sync';
 import Collections from '../resources/collections/sync';
 import DataObjects from '../resources/data-objects/sync';
 import Endpoints from '../resources/endpoints/sync';
+import Jobs from '../resources/jobs/sync';
 
-const map = [Namespaces, Collections, DataObjects, Endpoints];
+const map = [Namespaces, Collections, DataObjects, Endpoints, Jobs];
 
 export interface SyncOptions {
   follow: boolean;
@@ -39,7 +40,7 @@ export default class Sync {
         'Most resources have a namespace, request different namespace. The default namespace is "default".',
       );
       sub.option('-f, --follow', 'Follow process and watch in forderground');
-      sub.option('-t, --trace', 'Request resource logs including stacktraces');
+      sub.option('-t, --trace', 'Follow process and watch in forderground but include full backtraces');
       sub.option('-l, --level <name>', 'Specify log level for the process (emerg,error,warning,info,debug)');
       sub.option('-s, --simulate', 'Simulate sync process (No changes are made)');
       sub.option('--abort-on-error', 'Abort process if an error occurs');